Update dashboard_contents.md
I just added "% of users with transactions disputed" as we will have to keep those users happy. 

Comments:

I was thinking about the "daily transactions volume and $", as it will change depending on the day of the week, I would suggest doing it weekly. If we want a more day to day analysis (which I agree), we should compare it the daily trans-num with the same weekday of the previous week (or more than one).

Also, in the dashboard I would like to see the mean weekly (we can do it monthly) volume/amount of transactions per user BY how old is the user. To see if the user is "trusting" in our network, allowing more transactions.